+++ history.md
@@ -1,7 +1,18 @@
+##0.8.0
+
+
+ - Times in correct timezone.
+ - Better error handling.
+ - Better callbacks
+
+##0.5.2
+
+ - Added `failed`, `success` and `pending` boolean values n callback result
+
 ##0.5.1
 
-Bug fixes for successful payments
+ - Bug fixes for successful payments
 
 ##0.5.0
 
-Initial Release
+ - Initial Release